[Bronze III] Title: 알고리즘 수업 - 알고리즘의 수행 시간 4, Time: 104 ms, Memory: 14232 KB -BaekjoonHub

+++ 백준/Bronze/24265. 알고리즘 수업 － 알고리즘의 수행 시간 4/알고리즘 수업 － 알고리즘의 수행 시간 4.java
@@ -7,8 +7,8 @@ class Main {
 	 
 	    BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 	    
-	    long n = Integer.parseInt(br.readLine());
-	    System.out.println(n*(n-1)/2);
+	    int n = Integer.parseInt(br.readLine());
+	    System.out.println((long) n*(n-1)/2);
 	    System.out.print(2);
 	    
 	}
